Separating UI and Game logic
