+++ leaderboard.json
@@ -1,1 +1,1 @@
-[["jk", 12], ["rg", 6], ["peat", 5], ["eddie", 4], ["louis", 3]]
+[["jk", 12], ["rg", 6], ["edmund Phelan", 6], ["peat", 5], ["eddie", 4]]
